test: move e2e goaway tests to goaway_test.go
